+++ src/dolphin/workflows/wrapped_phase.py
@@ -48,7 +48,7 @@ def run(
     """
     if tqdm_kwargs is None:
         tqdm_kwargs = {}
-    logger = get_log(debug=debug)
+    logger = get_log(name=__name__, debug=debug)
     work_dir = cfg.work_directory
     logger.info("Running wrapped phase estimation in %s", work_dir)
 
@@ -322,6 +322,12 @@ def create_ifgs(
         )
         raise NotImplementedError(msg)
 
+    # Dedupe, in case different options made the same ifg
+    requested_ifgs = set(ifg_file_list)
+    # remove ones we aren't using (in the case of a single index)
+    written_ifgs = set(ifg_dir.glob("*.int*"))
+    for p in written_ifgs - requested_ifgs:
+        p.unlink()
     return ifg_file_list