+++ CHANGELOG.md
@@ -1,4 +1,26 @@
-# [Unreleased](https://github.com/isce-framework/dolphin/compare/v0.13.0...main)
+# [Unreleased](https://github.com/isce-framework/dolphin/compare/v0.15.0...main)
+
+# [0.15.0](https://github.com/isce-framework/dolphin/compare/v0.14.1...0.15.0) - 2024-02-16
+
+**Changed**
+
+- Combine the nodata region with the `mask_file` to pass through to unwrappers
+- Update regions which are nodata in interferograms to be nodata in unwrapped phase
+- Use `uint16` data type for connected component labels
+
+**Fixed**
+
+- Intersection of nodata regions for SLC stack are now all set to `nan` during phase linking
+
+# [0.14.1](https://github.com/isce-framework/dolphin/compare/v0.14.0...0.14.1) - 2024-02-15
+
+**Fixed**
+
+- Changed snaphu-py tile defaults to avoid max secondary arcs error in #233
+- Fixed `linalg.norm`` to be pixelwise in `process_coherence_matrices` in #234
+
+
+# [0.14.0](https://github.com/isce-framework/dolphin/compare/v0.13.0...0.14.0) - 2024-02-13
 
 **Fixed**
 - Temporal coherence and eigenvalue rasters were switched in their naming